Fixed compilation issues

+++ models.go
@@ -3,13 +3,13 @@ package main
 type User struct {
 	UserID         string  `json:"user_id"`
 	Name           string  `json:"name"`
-	AccountBalance float64 `json:"account_balance"`
+	AccountBalance float32 `json:"account_balance"`
 }
 
 type Stock struct {
 	Symbol       string  `json:"symbol"`
 	Name         string  `json:"name"`
-	CurrentPrice float64 `json:"current_price"`
+	CurrentPrice float32 `json:"current_price"`
 }
 
 type Trade struct {
@@ -18,7 +18,7 @@ type Trade struct {
 	Symbol   string  `json:"symbol"`
 	Quantity int     `json:"quantity"`
 	Type     string  `json:"type"` // buy/sell
-	Price    float64 `json:"price"`
+	Price    float32 `json:"price"`
 }
 
 type Holding struct {
